requirement.txt
groq==0.28.0
python-dotenv==1.1.0
fastapi==0.115.13
uvicorn==0.34.3